T1: (failed) optimization attempt for low bit planes in refpass (#932)
In the lowest bit plane in 6 to 8% of the cases, the 4 samples in a column
in the zig-zag pattern are in context 16 of the refpass. So let's add a
dedicated function for this.

However the (best) timings show that it is actually slower. With this special
case: 50205 ms vs 49570 ms originally on MAPA_005.jp2 reencoded with default options.